requirements.txt
djangorestframework==3.10.3
django==2.2.10
django-jet==1.0.8
sorl-thumbnail==12.5.0
psycopg2-binary==2.7.5
sorl-thumbnail==12.5.0
django_filter==2.2.0
supervisor==4.1.0
transliterate==1.10.2
gunicorn==20.0.4
whitenoise==5.0.1
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
Celery==4.4.0
redis==3.4.1
Pillow==6.2.2
django-widget-tweaks==1.4.5
easy-thumbnails==2.7
codecov
django-mptt==0.11.0
phonenumbers==8.12.1
django-phonenumber-field==4.0.0
django-config-models==2.0.0
django_extensions==2.2.9
drf-yasg==1.17.1
django-cors-headers==3.3.0
termcolor==1.1.0
paramiko==2.7.1
schedule==0.6.0